fix: issues with interface
