+++ package.json
@@ -112,6 +112,7 @@
     "vite": "6.3.5",
     "vite-plugin-checker": "0.9.3",
     "vite-plugin-istanbul": "7.0.0",
+    "vite-tsconfig-paths": "5.1.4",
     "vitest": "3.2.3"
   },
   "browserslist": {
