+++ src/constants.ts
@@ -1,4 +1,7 @@
+import { GPTVersion } from '@graasp/sdk';
+
 export const DEFAULT_BOT_USERNAME = 'Graasp Bot';
+export const DEFAULT_MODEL_VERSION = GPTVersion.GPT_3_5_TURBO;
 export const ANONYMOUS_USER = 'Anonymous';
 
 // UI
