+++ src/langs/en.json
@@ -24,9 +24,10 @@
     "CHATBOT_NAME_LABEL": "Chatbot Name",
     "CHATBOT_NAME_HELPER": "This is the name that users will see when they interact with the chatbot",
     "CHATBOT_NAME_DEFAULT_MESSAGE": "by default",
+    "CHATBOT_VERSION_DEFAULT_MESSAGE": "by default",
     "CHATBOT_PROMPT_LABEL": "Chatbot Prompt",
     "CHATBOT_PROMPT_HELPER_LABEL": "In-depth technical description",
-    "CHATBOT_PROMPT_HELPER": "This is the name that users will see when they interact with the chatbot",
+    "CHATBOT_PROMPT_HELPER": "This is a collection of examples of interactions that help the chatbot use an appropriate voice with users.",
     "CHATBOT_PROMPT_FORMAT_HELPER": "To describe the initial situation, create an object at the start of the array with 2 keys: 'role' and 'content'. For the initial description the role must be 'system'. Place your description of the initial situation in the 'content' key. Add interaction examples after that. Add one object per message with the role corresponding to either 'assistant' or 'user'.",
     "CHATBOT_PROMPT_FORMAT_EXAMPLE": "Here is an example",
     "CHATBOT_PROMPT_API_REFERENCE": "See the API reference",
@@ -66,7 +67,7 @@
     "GPT_3_5_TURBO_DESCRIPTION": "Optimized for speed and efficiency, suitable for applications where quick interactions are needed.",
     "GPT_4_DESCRIPTION": "Features deeper learning and more comprehensive knowledge, ideal for content creation and technical support.",
     "GPT_4_TURBO_DESCRIPTION": "Combines the depth of GPT-4 with enhanced speed, perfect for high-demand scenarios.",
-    "CHATBOT_MODEL_VERSION_HELPER": "This defines gpt version used within the model, default version is gpt-3.5-turbo-0125.",
+    "CHATBOT_MODEL_VERSION_HELPER": "This defines the  model that will be used. The default is: {{default}}.",
     "RECOMMENDED": "Recommended"
   }
 }
