+++ src/modules/settings/ChatbotSettings.tsx
@@ -2,66 +2,29 @@ import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import {
-  Accordion,
-  AccordionDetails,
-  AccordionSummary,
   Alert,
   Box,
   Button,
   Card,
   CardContent,
-  Chip,
   FormLabel,
-  Link,
-  MenuItem,
-  Select,
   Stack,
-  TextField,
-  TextareaAutosize,
   Typography,
-  styled,
 } from '@mui/material';
 
-import { ChatBotMessage, GPTVersion } from '@graasp/sdk';
-
-import { ChevronDown, Edit, Undo2 } from 'lucide-react';
+import { Edit, Undo2 } from 'lucide-react';
 
 import { ChatbotPromptSettings, SettingsKeys } from '@/config/appSetting';
 import { hooks, mutations } from '@/config/queryClient';
-import { SETTING_CHATBOT_PROMPT_CODE_EDITOR_CY } from '@/config/selectors';
-import { DEFAULT_BOT_USERNAME, SMALL_BORDER_RADIUS } from '@/constants';
-import { showErrorToast } from '@/utils/toast';
+import { DEFAULT_BOT_USERNAME, DEFAULT_MODEL_VERSION } from '@/constants';
 
 import CodeEditor from '../common/CodeEditor';
-
-const TextArea = styled(TextareaAutosize)(({ theme }) => ({
-  borderRadius: SMALL_BORDER_RADIUS,
-  padding: theme.spacing(2),
-  fontSize: '1rem',
-  boxSizing: 'border-box',
-  resize: 'vertical',
-  border: 0,
-  outline: 'solid rgba(80, 80, 210, 0.5) 1px',
-  // make sure the outline is offset by the same amount that it is wide to not overflow
-  outlineOffset: '-1px',
-  width: '100%',
-  minWidth: '0',
-  minHeight: `calc(1rem + 2*${theme.spacing(2)})`,
-  transition: 'outline 250ms ease-in-out',
-  '&:focus': {
-    outline: 'solid var(--graasp-primary) 2px !important',
-  },
-  '&:hover': {
-    outline: 'solid var(--graasp-primary) 1px ',
-  },
-}));
+import { ChatbotEditionView } from './chatbot/ChatbotEditingView';
 
 const ChatbotSettings = (): JSX.Element => {
   const { t } = useTranslation();
   const { mutate: postSetting } = mutations.usePostAppSetting();
   const { mutate: patchSetting } = mutations.usePatchAppSetting();
-  const [unsavedChanges, setUnsavedChanges] = useState(false);
-  const [formattingError, setFormattingError] = useState(false);
   const [isEditing, setIsEditing] = useState(false);
   const { data: chatbotPromptSettings } =
     hooks.useAppSettings<ChatbotPromptSettings>({
@@ -73,98 +36,31 @@ const ChatbotSettings = (): JSX.Element => {
   const chatbotCue = chatbotPrompt?.data?.chatbotCue ?? '';
   const chatbotName = chatbotPrompt?.data?.chatbotName ?? DEFAULT_BOT_USERNAME;
   const chatbotVersion =
-    chatbotPrompt?.data?.gptVersion ?? GPTVersion.GPT_3_5_TURBO;
-
-  const [newChatbotPrompt, setNewChatbotPrompt] = useState(
-    stringifiedJsonPrompt,
-  );
-  const [newChatbotCue, setNewChatbotCue] = useState(chatbotCue);
-  const [newChatbotName, setNewChatbotName] = useState(chatbotName);
-  const [newChatbotVersion, setNewChatbotVersion] = useState(chatbotVersion);
+    chatbotPrompt?.data?.gptVersion ?? DEFAULT_MODEL_VERSION;
 
   const doneEditing = (): void => {
     setIsEditing(false);
   };
 
-  const validatePrompt = <T,>(
-    prompt: string,
-    callbacks: {
-      onSuccess?: () => void;
-      onError?: () => void;
-    },
-  ): T | undefined => {
-    try {
-      const jsonNewChatbotPrompt = JSON.parse(prompt);
-      callbacks.onSuccess?.();
-      return jsonNewChatbotPrompt;
-    } catch {
-      callbacks.onError?.();
+  const handleOnSave = (data: ChatbotPromptSettings): void => {
+    // setting does not exist
+    if (!chatbotPrompt) {
+      postSetting({
+        data,
+        name: SettingsKeys.ChatbotPrompt,
+      });
+    } else {
+      patchSetting({
+        id: chatbotPrompt.id,
+        data,
+      });
     }
-    return undefined;
-  };
-
-  const hasNoFormattingErrors = (): void => setFormattingError(false);
-  const hasFormattingErrors = (): void => setFormattingError(true);
-
-  const handleChangeChatbotPrompt = (value: string): void => {
-    validatePrompt(value, {
-      onSuccess: hasNoFormattingErrors,
-      onError: hasFormattingErrors,
-    });
-    setNewChatbotPrompt(value);
-    setUnsavedChanges(true);
-  };
-
-  const handleChangeChatbotCue = (value: string): void => {
-    setNewChatbotCue(value);
-    setUnsavedChanges(true);
-  };
-
-  const handleChangeChatbotName = (value: string): void => {
-    setNewChatbotName(value);
-    setUnsavedChanges(true);
-  };
-
-  const handleCancel = (): void => {
-    // reset fields
-    setNewChatbotPrompt(stringifiedJsonPrompt);
-    setNewChatbotName(chatbotName);
-    setNewChatbotCue(chatbotCue);
-    // resume editing
+    // close the editing view
     doneEditing();
   };
 
-  const handleSave = (): void => {
-    const jsonNewChatbotPrompt = validatePrompt<ChatBotMessage[]>(
-      newChatbotPrompt,
-      {
-        onError: () => {
-          hasFormattingErrors();
-          showErrorToast(t('ERROR_PROMPT_NOT_IN_JSON_FORMAT'));
-        },
-      },
-    );
-    if (jsonNewChatbotPrompt) {
-      const data: ChatbotPromptSettings = {
-        initialPrompt: jsonNewChatbotPrompt,
-        chatbotCue: newChatbotCue,
-        chatbotName: newChatbotName,
-        gptVersion: newChatbotVersion,
-      };
-      // setting does not exist
-      if (!chatbotPrompt) {
-        postSetting({
-          data,
-          name: SettingsKeys.ChatbotPrompt,
-        });
-      } else {
-        patchSetting({
-          id: chatbotPrompt.id,
-          data,
-        });
-      }
-      doneEditing();
-    }
+  const handleCancel = (): void => {
+    doneEditing();
   };
 
   return (
@@ -196,154 +92,15 @@ const ChatbotSettings = (): JSX.Element => {
       )}
 
       {isEditing ? (
-        <Stack spacing={2}>
-          <Box>
-            <Stack>
-              <FormLabel>{t('CHATBOT_NAME_LABEL')}</FormLabel>
-              <Typography variant="caption" color="text.secondary">
-                {t('CHATBOT_NAME_HELPER')}
-              </Typography>
-            </Stack>
-            <TextField
-              fullWidth
-              id={SETTING_CHATBOT_PROMPT_CODE_EDITOR_CY}
-              value={newChatbotName}
-              onChange={({ target: { value } }) =>
-                handleChangeChatbotName(value)
-              }
-            />
-          </Box>
-          <Box>
-            <Stack spacing={2}>
-              <Stack>
-                <FormLabel>{t('CHATBOT_PROMPT_LABEL')}</FormLabel>
-                <Typography variant="caption" color="text.secondary">
-                  {t('CHATBOT_PROMPT_HELPER')}
-                </Typography>
-                <CodeEditor
-                  value={newChatbotPrompt}
-                  onChange={(value: string) => handleChangeChatbotPrompt(value)}
-                />
-                {formattingError ? (
-                  <Typography color="error">
-                    {t('ERROR_PROMPT_NOT_IN_JSON_FORMAT')}
-                  </Typography>
-                ) : null}
-              </Stack>
-              <Accordion disableGutters>
-                <AccordionSummary
-                  expandIcon={<ChevronDown />}
-                  aria-controls="panel1-content"
-                  id="help-me"
-                >
-                  {t('CHATBOT_PROMPT_HELPER_LABEL')}
-                </AccordionSummary>
-                <AccordionDetails>
-                  <Stack spacing={1}>
-                    <Typography variant="caption" color="text.secondary">
-                      {t('CHATBOT_PROMPT_FORMAT_HELPER')}
-                    </Typography>
-                    <Typography variant="caption" color="text.secondary">
-                      {t('CHATBOT_PROMPT_FORMAT_EXAMPLE')}
-                    </Typography>
-                    <CodeEditor
-                      value={JSON.stringify(
-                        [
-                          {
-                            role: 'system',
-                            content: 'You are a helpful assistant.',
-                          },
-                          {
-                            role: 'user',
-                            content: 'Who won the world series in 2020?',
-                          },
-                          {
-                            role: 'assistant',
-                            content:
-                              'The Los Angeles Dodgers won the World Series in 2020.',
-                          },
-                          { role: 'user', content: 'Where was it played?' },
-                        ],
-                        null,
-                        2,
-                      )}
-                      readOnly
-                      fontSize="10px"
-                    />
-
-                    <Link
-                      variant="caption"
-                      target="_blank"
-                      href="https://platform.openai.com/docs/api-reference/chat/create#chat-create-messages"
-                    >
-                      {t('CHATBOT_PROMPT_API_REFERENCE')}
-                    </Link>
-                  </Stack>
-                </AccordionDetails>
-              </Accordion>
-            </Stack>
-          </Box>
-          <Box>
-            <Stack>
-              <FormLabel>{t('MODEL_VERSION')}</FormLabel>
-              <Typography variant="caption" color="text.secondary">
-                {t('CHATBOT_MODEL_VERSION_HELPER')}
-              </Typography>
-            </Stack>
-            <Select
-              value={newChatbotVersion}
-              label={t('MODEL_VERSION')}
-              onChange={(event) => {
-                setNewChatbotVersion(event.target.value as GPTVersion);
-                setUnsavedChanges(true);
-              }}
-              fullWidth
-              renderValue={(selected) => <Typography>{selected}</Typography>}
-            >
-              {Object.entries(GPTVersion).map(([key, v]) => (
-                <MenuItem key={v} value={v} sx={{ display: 'block' }}>
-                  <Stack direction="row" spacing={1}>
-                    <Typography>{v}</Typography>
-                    {v === GPTVersion.GPT_3_5_TURBO && (
-                      <Chip
-                        label={t('RECOMMENDED')}
-                        size="small"
-                        color="primary"
-                      />
-                    )}
-                  </Stack>
-                  <Typography variant="caption" color="text.secondary">
-                    {t(`${key}_DESCRIPTION`)}
-                  </Typography>
-                </MenuItem>
-              ))}
-            </Select>
-          </Box>
-          <Box>
-            <Stack>
-              <FormLabel>{t('CHATBOT_CUE_LABEL')}</FormLabel>
-              <Typography variant="caption" color="text.secondary">
-                {t('CHATBOT_CUE_HELPER')}
-              </Typography>
-            </Stack>
-            <TextArea
-              id={SETTING_CHATBOT_PROMPT_CODE_EDITOR_CY}
-              value={newChatbotCue}
-              onChange={({ target: { value } }) =>
-                handleChangeChatbotCue(value)
-              }
-            />
-          </Box>
-          <Box alignSelf="flex-end">
-            <Button
-              onClick={handleSave}
-              disabled={!unsavedChanges}
-              variant="outlined"
-            >
-              {unsavedChanges ? t('SAVE_LABEL') : t('SAVED_LABEL')}
-            </Button>
-          </Box>
-        </Stack>
+        <ChatbotEditionView
+          initialValue={{
+            name: chatbotName,
+            cue: chatbotCue,
+            version: chatbotVersion,
+            prompt: stringifiedJsonPrompt,
+          }}
+          onSave={handleOnSave}
+        />
       ) : (
         <Card elevation={0} variant="outlined">
           <CardContent sx={{ pb: 0 }}>
@@ -365,6 +122,22 @@ const ChatbotSettings = (): JSX.Element => {
                 </Typography>
               </Stack>
               <Stack direction="column">
+                <Stack direction="row" spacing={1}>
+                  <FormLabel> {t('MODEL_VERSION')}:</FormLabel>
+                  <Typography>{chatbotVersion}</Typography>
+                  {chatbotVersion === DEFAULT_MODEL_VERSION ? (
+                    <Typography color="text.disabled">
+                      ({t('CHATBOT_VERSION_DEFAULT_MESSAGE')})
+                    </Typography>
+                  ) : null}
+                </Stack>
+                <Typography variant="caption" color="text.secondary">
+                  {t('CHATBOT_MODEL_VERSION_HELPER', {
+                    default: DEFAULT_MODEL_VERSION,
+                  })}
+                </Typography>
+              </Stack>
+              <Stack direction="column">
                 <FormLabel>{t('CHATBOT_PROMPT_LABEL')}:</FormLabel>
                 <Typography variant="caption" color="text.secondary">
                   {t('CHATBOT_PROMPT_HELPER')}
@@ -372,28 +145,6 @@ const ChatbotSettings = (): JSX.Element => {
                 <Box width="100%">
                   <CodeEditor value={stringifiedJsonPrompt} readOnly />
                 </Box>
-              </Stack>
-              <Stack direction="column">
-                <Stack>
-                  <FormLabel>{t('CHATBOT_CUE_LABEL')}:</FormLabel>
-                  <Typography variant="caption" color="text.secondary">
-                    {t('CHATBOT_CUE_HELPER')}
-                  </Typography>
-                </Stack>
-                {chatbotCue ? (
-                  <Typography>{chatbotCue}</Typography>
-                ) : (
-                  <Typography color="text.disabled" fontStyle="italic">
-                    {t('CHATBOT_CUE_EMPTY_MESSAGE')}
-                  </Typography>
-                )}
-              </Stack>
-              <Stack direction="column">
-                <FormLabel> {t('MODEL_VERSION')}</FormLabel>
-                <Typography variant="caption" color="text.secondary">
-                  {t('CHATBOT_MODEL_VERSION_HELPER')}
-                </Typography>
-                <Typography>{newChatbotVersion}</Typography>
               </Stack>
               <Stack direction="column">
                 <Stack>
