+++ src/modules/settings/SettingsView.tsx
@@ -1,12 +1,12 @@
 import { Stack } from '@mui/material';
 
 import ChatbotSettings from './ChatbotSettings';
-import GeneralSettingsView from './GeneralSettingsView';
+import { GeneralSettingsView } from './GeneralSettingsView';
 
 const SettingsView = (): JSX.Element => (
   <Stack spacing={1}>
-    <GeneralSettingsView />
     <ChatbotSettings />
+    <GeneralSettingsView />
   </Stack>
 );
 export default SettingsView;
