+++ tsconfig.json
@@ -2,12 +2,7 @@
   "compilerOptions": {
     "target": "ESNext",
     "useDefineForClassFields": true,
-    "lib": [
-      "DOM",
-      "DOM.Iterable",
-      "ESNext",
-      "ES2021.String"
-    ],
+    "lib": ["DOM", "DOM.Iterable", "ESNext", "ES2021.String"],
     "allowJs": false,
     "skipLibCheck": true,
     "esModuleInterop": true,
@@ -20,26 +15,14 @@
     "isolatedModules": true,
     "noEmit": true,
     "jsx": "react-jsx",
-    "typeRoots": [
-      "node_modules",
-      "node_modules/@types",
-      "./src/@types"
-    ],
-    "types": [
-      "vite/client",
-      "cypress"
-    ],
+    "typeRoots": ["node_modules", "node_modules/@types", "./src/@types"],
+    "types": ["vite/client", "cypress"],
     "baseUrl": "./src",
     "paths": {
-      "@/*": [
-        "./*"
-      ]
+      "@/*": ["./*"]
     }
   },
-  "include": [
-    "src",
-    "cypress"
-  ],
+  "include": ["src", "cypress"],
   "references": [
     {
       "path": "./tsconfig.node.json"
